cleaner/requirements.txt
duckdb==1.1.3
polars-lts-cpu
pyarrow
minio==7.2.7
pika==1.3.2
prometheus-client==0.20.0
